Refactor routes and polish mobile/profile flows
