+++ app/routes/landing.py
@@ -1,0 +1,8 @@
+from flask import render_template
+
+from ..blueprints import main_bp
+
+
+@main_bp.route("/")
+def landing():
+    return render_template("landing.html")
