+++ app/static/css/core/layout.css
@@ -370,3 +370,66 @@
     color: #aab3d9;
 }
 
+@media (max-width: 900px) {
+    .app-shell {
+        display: block;
+        padding: 16px;
+        gap: 0;
+    }
+
+    .sidebar {
+        display: none;
+    }
+
+    .main-content {
+        border-radius: 20px;
+        padding: 18px;
+        padding-bottom: calc(92px + env(safe-area-inset-bottom));
+    }
+
+    .mobile-bottom-nav {
+        position: fixed;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        z-index: 1000;
+
+        display: grid;
+        grid-template-columns: repeat(7, minmax(0, 1fr));
+        align-items: center;
+        gap: 4px;
+
+        background: rgba(9, 14, 31, 0.98);
+        border-top: 1px solid rgba(139, 92, 246, 0.28);
+        backdrop-filter: blur(12px);
+
+        padding: 8px 8px calc(8px + env(safe-area-inset-bottom));
+    }
+
+    .mobile-bottom-nav a {
+        text-decoration: none;
+        color: #b8c2dc;
+
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        gap: 4px;
+
+        min-height: 56px;
+        border-radius: 12px;
+        font-size: 11px;
+        font-weight: 600;
+    }
+
+    .mobile-bottom-nav a i {
+        font-size: 18px;
+        line-height: 1;
+    }
+
+    .mobile-bottom-nav a.active {
+        color: #ffffff;
+        background: linear-gradient(180deg, rgba(168, 85, 247, 0.22), rgba(59, 130, 246, 0.16));
+        box-shadow: inset 0 0 0 1px rgba(192, 132, 252, 0.22);
+    }
+}
