+++ app/templates/forgot_password.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Forgot Password | StudySync</title>
+    <title>Forgot Password | StudyCollabz</title>
 
         <link rel="stylesheet" href="{{ url_for('static', filename='css/core/base.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/core/layout.css') }}">
@@ -133,7 +133,6 @@
 
 </body>
 </html>
-
