+++ app/templates/landing.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>StudySync | Made for UWA Students</title>
+    <title>StudyCollabz | Made for UWA Students</title>
         <link rel="stylesheet" href="{{ url_for('static', filename='css/core/base.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/core/layout.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/components/buttons.css') }}">
@@ -13,7 +13,7 @@
 <body class="landing-page">
     <header class="landing-nav" aria-label="Main navigation">
         <a class="landing-logo" href="{{ url_for('main.landing') }}">
-            Study<span>Sync</span>
+            Study<span>Collabz</span>
         </a>
 
         <nav class="landing-actions" aria-label="Account actions">
@@ -33,7 +33,7 @@
                 </h1>
 
                 <p class="landing-text">
-                    StudySync helps UWA students find study partners by matching unit codes,
+                    StudyCollabz helps UWA students find study partners by matching unit codes,
                     weekly availability, preferred study mode, and group size.
                 </p>
 
@@ -114,7 +114,7 @@
                     </footer>
                 </article>
 
-                <div class="unit-stack-links" aria-label="StudySync benefits">
+                <div class="unit-stack-links" aria-label="StudyCollabz benefits">
                     <span>Find students</span>
                     <span>Match by units</span>
                     <span>Study together</span>
@@ -122,7 +122,7 @@
             </section>
         </section>
 
-        <section class="landing-feature-grid" aria-label="StudySync features">
+        <section class="landing-feature-grid" aria-label="StudyCollabz features">
             <article class="landing-feature-card">
                 <span class="feature-mark" aria-hidden="true">01</span>
                 <h2>Match by unit</h2>
@@ -144,5 +144,4 @@
     </main>
 </body>
 </html>
-
