+++ app/templates/loginpage.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login | StudySync</title>
+    <title>Login | StudyCollabz</title>
 
         <link rel="stylesheet" href="{{ url_for('static', filename='css/core/base.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/core/layout.css') }}">
@@ -16,7 +16,7 @@
     <main class="login-wrapper">
 
         <section class="login-left">
-            <h2>StudySync</h2>
+            <h2>StudyCollabz</h2>
 
             <h1>
                 Welcome back,<br>
