+++ app/templates/notifications.html
@@ -1,6 +1,6 @@
 {% extends "base.html" %}
 
-{% block title %}Notifications – StudySync{% endblock %}
+{% block title %}Notifications – StudyCollabz{% endblock %}
 
 {% block head %}
 <link rel="stylesheet" href="{{ url_for('static', filename='css/pages/notifications.css') }}">
