+++ app/templates/onboarding.html
@@ -9,7 +9,7 @@ Onboarding{% endblock %}
 <div class="onboarding-wrapper">
 <div class="page-header">
     <div>
-        <h1>Welcome to StudySync</h1>
+        <h1>Welcome to StudyCollabz</h1>
         <p>Quick setup in {{ total_steps }} steps.</p>
     </div>
 </div>
@@ -264,4 +264,3 @@ unitInputs.forEach((input) => {
 </script>
 </div>
 {% endblock %}
-
