+++ app/templates/reset_password.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Reset Password | StudySync</title>
+    <title>Reset Password | StudyCollabz</title>
 
     <link rel="stylesheet" href="{{ url_for('static', filename='css/pages/auth.css') }}">
 </head>
@@ -17,7 +17,7 @@
             <h1>Reset Your Password</h1>
 
             <p>
-                Enter your new password below to secure your StudySync account.
+                Enter your new password below to secure your StudyCollabz account.
             </p>
 
             <form method="POST" class="reset-form">
@@ -228,7 +228,6 @@
 
 </body>
 </html>
-
