+++ app/templates/signup.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Register | StudySync</title>
+    <title>Register | StudyCollabz</title>
 
         <link rel="stylesheet" href="{{ url_for('static', filename='css/core/base.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/core/layout.css') }}">
@@ -17,12 +17,12 @@
 
         <section class="register-left">
 
-            <h2>Study<span>Sync</span></h2>
+            <h2>Study<span>Collabz</span></h2>
 
             <h1>Register</h1>
 
             <p class="register-subtitle">
-                Create your StudySync account
+                Create your StudyCollabz account
             </p>
 
             <p class="register-description">
