+++ app/templates/userpages.html
@@ -186,7 +186,7 @@
                 {% if current_user %}
                 {{ current_user.first_name }} {{ current_user.last_name }}
                 {% else %}
-                StudySync User
+                StudyCollabz User
                 {% endif %}
 
             </div>
